[JENKINS-55423] Support custom Jenkins Root URLs for webhook setup (#165)
In some cases, the Jenkins server can be seen to its clients (here the Bitbucket Cloud or Server instance) under a different URL than it declares in its local networking segment. This PR adds an optional configurable setting for each endpoint to use a custom Jenkins Root URL when managing webhooks for that endpoint.